test-requirements.txt
bashate>=0.2 # Apache-2.0
coverage>=4.0  # Apache-2.0
doc8  # Apache-2.0
hacking>=0.12.0,<0.13  # Apache-2.0
openstackdocstheme>=1.5.0  # Apache-2.0
sphinx>=1.6.2 # BSD
testrepository>=0.0.18  # Apache-2.0/BSD
python-openstackclient>=3.16.1
python-designateclient
python-octaviaclient
python-senlinclient

# Python 2.7 dependencies
mock; python_version < '3.0'
unittest2; python_version < '3.0'
